Use the itme specific base role arn
We need a general soltuion for passing variables from terraform to
helfile. In the past I have just created a file from terraform that is
meant to be sourced before running helmfile apply. That probably
requires setting up CD for the cluster, which is an effort I don't want
to deal with yet.